==> Lessons/09_DataVisualization.Rmd <==
---
title: "9: Data Visualization"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2019"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Perform simple data visualizations in the R package `ggplot`
2. Develop skills to adjust aesthetics and layers in graphs
3. Apply a decision tree framework for appropriate graphing methods

## SET UP YOUR DATA ANALYSIS SESSION

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse)

PeterPaul.chem.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv")
PeterPaul.nutrients.gathered <- read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv")
PeterPaul.chem.nutrients.summaries <- read.csv("./Data/Processed/NTL-LTER_Lake_Summaries_PeterPaul_Processed.csv")
EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1718_Processed.csv")

EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")

```

## GGPLOT

ggplot, called from the package `ggplot2`, is a graphing and image generation tool in R. This package is part of tidyverse. While base R has graphing capabilities, ggplot has the capacity for a wider range and more sophisticated options for graphing. 
ggplot has only a few rules: 

* The first line of ggplot code always starts with `ggplot()`
* A data frame must be specified within the `ggplot()` function. Additional datasets can be specified in subsequent layers.
* Aesthetics must be specified, most commonly x and y variables but including others. Aesthetics can be specified in the `ggplot()` function or in subsequent layers.
* Additional layers must be specified to fill the plot. 

### Geoms

Here are some commonly used layers for plotting in ggplot: 

* geom_bar
* geom_histogram
* geom_freqpoly
* geom_boxplot
* geom_violin
* geom_dotplot
* geom_point
* geom_errorbar
* geom_smooth
* geom_line
* geom_area
* geom_abline (plus geom_hline and geom_vline)
* geom_text

### Aesthetics
Here are some commonly used aesthetic types that can be manipulated in ggplot: 

* color
* fill
* shape
* size
* transparency

### Plotting continuous variables over time: Scatterplot
```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot: the outcome of the graph shows a seasonal graph, not good to make lines over the pattern
ggplot(EPAair, aes(x = Date, y = Ozone)) + 
  geom_point()

O3plot <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone))
print(O3plot)  # make it show your graph

# Fix this code
O3plot2 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone), color = "blue")
print(O3plot2)

#What if we make this one? not a good one, coz it's redundant
O3plot2 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone, color = Ozone))
print(O3plot2)

# Add additional variables
# we put as.factor for year coz year is in integer right now
#the output is a pretty busy graph, a lot of things going on, but 
PMplot <- 
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year), color = Site.Name)) +
  geom_point()
print(PMplot)

# Separate plot with facets, basically separate them into three components
PMplot.faceted <-
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year))) +
  geom_point() +
  facet_wrap(vars(Site.Name), nrow = 3) #nrow =3 means to stack the 3 components into 3 rows, you can do 3 columns too
print(PMplot.faceted)

PMplot.faceted2 <-
  ggplot(EPAair, aes(x = Month, y = PM2.5)) +
  geom_point() +
  facet_grid(Site.Name ~ Year)  # create it by site by year
print(PMplot.faceted2)

# Filter dataset within plot building
#Subset returns subsets of vectors, matrices or data frames which meet conditions.
O3plot.Blackstone <- 
  ggplot(subset(EPAair, Site.Name == "Blackstone"), aes(x = Date, y = Ozone)) + 
  geom_point() +
  geom_line() #connect the dots with a line; consider whether this is necessary
print(O3plot.Blackstone)

# Exercise: build your own scatterplots of PeterPaul.chem.nutrients

# 1. 
# Plot surface temperatures by day of year. 
# Color your points by year, and facet by lake in two rows.

pepa.chem <- 
  ggplot(subset(PeterPaul.chem.nutrients, depth = 0), aes(x = daynum, y = temperature_C, color = year4)) + 
  geom_point() +
  facet_wrap(vars(lakename))
print(pepa.chem)


#2. 
# Plot temperature by date. Color your points by depth.
# Change the size of your point to 0.5

pepa.chem1 <- ggplot(PeterPaul.chem.nutrients, aes(x = sampledate, y = temperature_C, color = depth)) +
geom_point(size = 0.5)
print(pepa.chem1)


```
### Plotting the relationship between two continuous variables: Scatterplot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
lightvsDO <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point()
print(lightvsDO)

# Adjust axes
lightvsDOfixed <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point() +
  xlim(0, 250) +
  ylim(0, 20)
print(lightvsDOfixed)

# Depth in the fields of limnology and oceanography is on a reverse scale
tempvsdepth <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth)) +
  #ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth, color = daynum)) +
  geom_point() +
  scale_y_reverse()
print(tempvsdepth)

NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth)) +
  geom_point() +
  geom_smooth(method = lm) + #if you just say geom_smooth(), the line will stay close to the data
  geom_abline(aes(slope = 16, intercept = 0)) #put lty = 2 "line type", you will get a dotted line, but put it outside the aes(); the graph also shows that phosphates area more limited, for every 1 p molecule there is 16 nitrogen molecule
print(NvsP)

# Exercise: Plot relationships between air quality measurements

# 1. 
# Plot AQI values for ozone by PM2.5, colored by site. 
# Add a line of best fit for the linear regression of these variables.


ozonepm25 <- 
  ggplot(EPAair, aes(x = PM2.5, y = Ozone, color = Site.Name)) +
  geom_point() +
  geom_smooth(method = lm) +
print(ozonepm25)

#if you wanna specify aes in your geom_point, make sure you just copy paste it in the rest of your other geom
# ozonepm25 <- 
#   ggplot(EPAair) +
#   geom_point(aes(x = PM2.5, y = Ozone, color = Site.Name)) +
#   geom_smooth(aes(x = PM2.5, y = Ozone, color = Site.Name), method = lm) +
# print(ozonepm25)
  
```

### Plotting continuous vs. categorical variables

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
 # Barplot + error bars
PeterPaul.nutrient.summaries <- PeterPaul.nutrients.gathered %>%
  group_by(lakename, nutrient)  %>% #if you wannait to be group by both lakename and nutrient you need to put them in the same groupby function
  summarise(sd = sd(concentration), 
            mean = mean(concentration))

Nutrientplot <- 
  ggplot(PeterPaul.nutrients.gathered) +
  geom_bar(aes(x = lakename, y = concentration, fill = as.factor(nutrient)), # why did we use fill? Because in bar, there is both fill and color, the color only defines the border so you need to specify what fills the color inside
           position = "dodge", stat = "summary", fun.y = "mean")             # what's happening here? Position meanns to not stack them on top of each other
print(Nutrientplot)

Nutrientplot2 <- 
  ggplot(PeterPaul.nutrient.summaries, aes(x = lakename, y = mean, fill = as.factor(nutrient))) + #
  geom_bar(stat = "identity", position = "dodge") + # what does the stat command do? Bar chart requires to define the stat, that's why we put "summary", "mean", something to put in the stat(); in this case, we already have sd and mean, which is what we need, so we only need to ask R to point to those numbers, which is "identity"
  geom_errorbar(aes(ymin = mean-sd, ymax = mean+sd), # how do we specify error bars? geom error bar needs aes too, it wants minimum and max vaue, in this case, we want the min to be mean-sd, it is already in the dataframe; so if you don't have, you need to create it
                 position = "dodge") # if you don't use position, it will all combine into one giant error bar
print(Nutrientplot2)
#overall, when we see the error bar, sometimes it goes to negative which does not make sense coz nutrient cannot go down than 0
#envi field is moving away from bar + error bar

# Are there more effective ways to produce summary stats for categories?

# Box and whiskers plot
Nutrientplot3 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_boxplot(aes(color = nutrient)) # Why didn't we use "fill"?
print(Nutrientplot3)
# this one is better since there is no negative value anymore. 


# Dot plot
Nutrientplot4 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_dotplot(aes(color = nutrient), binaxis = "y", binwidth = 1, 
               stackdir = "center", position = "dodge") #
print(Nutrientplot4)

# Violin plot
Nutrientplot5 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_violin(aes(color = nutrient)) #
print(Nutrientplot5)

# Frequency polygons
# Using a tidy dataset
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients) +
  geom_freqpoly(aes(x = tn_ug), color = "black") +
  geom_freqpoly(aes(x = tp_ug), color = "darkblue") +
  geom_freqpoly(aes(x = nh34), color = "darkgray") +
  geom_freqpoly(aes(x = no23), color = "gray") +
  geom_freqpoly(aes(x = po4), color = "blue") 
print(Nutrientplot6)

# Using a gathered dataset
Nutrientplot7 <-   
  ggplot(PeterPaul.nutrients.gathered) +
  geom_freqpoly(aes(x = concentration, color = nutrient))
print(Nutrientplot7)

# Exercise: Plot distributions of AQI values for EPAair

# 1. 
# Create a bar chart plus standard deviation error bars for PM2.5, divided by year. 
# Create separate bars for each site. 

gather.pm25errorbar <- EPAair %>%
  group_by(Site.Name, Year) %>% 
  summarise(sd = sd(PM2.5, na.rm = TRUE),  # na.rm means remove all the NAs
            mean = mean(PM2.5, na.rm = TRUE))
  

pm25errorbar <-
  ggplot(gather.pm25errorbar, aes(x=as.factor(Year), y=mean, fill= Site.Name)) +
  geom_bar(position = "dodge", stat = "identity") +
  geom_errorbar(aes(ymin = mean-sd, ymax = mean + sd), position = "dodge")
print(pm25errorbar)


# 2. 
# Create a new plot that better depicts the distribution of PM2.5 concentrations. 
# Divide your graph by year and site.

```
